Adding additional ccomponenet for Posting Comments. Rerender needs to be adjusted
